Further localize jax samplers
