+++ netket/sampler/__init__.py
@@ -8,3 +8,8 @@ from .metropolis_hamiltonian import *
 from .custom_sampler import *
 
 from .exact_sampler import *
+
+from ..utils import jax_available
+
+if jax_available:
+    from . import jax
