+++ netket/sampler/_kernels.py
@@ -23,30 +23,3 @@ def _CustomKernel(machine, move_operators, move_weights=None):
 @singledispatch
 def _HamiltonianKernel(machine, hamiltonian):
     return numpy._HamiltonianKernel(hamiltonian)
-
-
-# Register Jax kernels here
-from ..utils import jax_available
-
-if jax_available:
-    from . import jax
-    from ..machine import Jax as JaxMachine
-
-    @_LocalKernel.register(JaxMachine)
-    def _Jax_LocalKernel(machine):
-        return jax._LocalKernel(machine.hilbert.local_states, machine.input_size)
-
-    @_ExchangeKernel.register(JaxMachine)
-    def _Jax_ExchangeKernel(machine, d_max):
-        return jax._ExchangeKernel(machine.hilbert, d_max)
-
-    @_HamiltonianKernel.register(JaxMachine)
-    def _Jax_HamiltonianKernel(machine, hamiltonian):
-        raise NotImplementedError
-
-    @_CustomKernel.register(JaxMachine)
-    def _Jax_CustomKernel(machine, move_operators, move_weights=None):
-        raise NotImplementedError
-
-
-# Register PyTorch kernels here
