+++ netket/sampler/metropolis_hastings.py
@@ -75,18 +75,3 @@ def MetropolisHastingsPt(machine, kernel, n_replicas=32, sweep_size=None):
 
     """
     return numpy.MetropolisHastingsPt(machine, kernel, n_replicas, sweep_size)
-
-
-from ..utils import jax_available
-
-if jax_available:
-    from . import jax
-    from ..machine import Jax as JaxMachine
-
-    @MetropolisHastings.register(JaxMachine)
-    def _JaxMetropolisHastings(machine, kernel, n_chains=16, sweep_size=None):
-        return jax.MetropolisHastings(machine, kernel, n_chains, sweep_size)
-
-    @MetropolisHastingsPt.register(JaxMachine)
-    def _JaxMetropolisHastingsPt(machine, kernel, n_replicas=32, sweep_size=None):
-        raise NotImplementedError
